feat: updated keyple to 0.9-alpha2 (Windows)

+++ component/keyple-calypso/src/main/command/po/builder/WriteRecordCmdBuild.cpp
@@ -67,7 +67,7 @@ bool WriteRecordCmdBuild::isSessionBufferUsed() const
     return true;
 }
 
-int WriteRecordCmdBuild::getSfi() const
+uint8_t WriteRecordCmdBuild::getSfi() const
 {
     return mSfi;
 }
